+++ paddle/fluid/pybind/CMakeLists.txt
@@ -42,7 +42,7 @@ set(PYBIND_DEPS
     pd_op_dialect
     program_translator
     pd_inplace_pass
-    conv2d_fuse_pass
+    fusion_passes
     pir
     new_profiler
     jit_layer
